Move EqST to test suite
