+++ cardano-crypto-tests/cardano-crypto-tests.cabal
@@ -54,6 +54,7 @@ library
                         Test.Crypto.VRF
                         Test.Crypto.Regressions
                         Test.Crypto.Instances
+                        Test.Crypto.EqST
                         Bench.Crypto.DSIGN
                         Bench.Crypto.VRF
                         Bench.Crypto.KES
@@ -86,6 +87,7 @@ library
                       , criterion
                       , base16-bytestring
                       , tasty-hunit
+                      , vector
 
   if flag(secp256k1-support)
     cpp-options: -DSECP256K1_ENABLED
